refactor: colocate header components
- move AppHeader/TeamSelector/HeaderNavLinks into header folder
- update import paths
